Reorganize public api
