+++ src/html/mod.rs
@@ -1,8 +1,9 @@
-mod builder;
 mod format;
 mod generator;
 mod escape;
+mod output;
+pub mod builders;
 
-pub use self::builder::{Builder, TagStartBuilder, Output};
+pub use self::output::*;
 pub use self::format::*;
 pub use self::generator::*;
